V0.9

+++ HRC Webapp/manual_edit.py
@@ -10,14 +10,14 @@ import sql
 
 def open_reports():
     try:
-        f = open("reports_record.txt","r")
+        f = open("reports_record.dat","r")
     except:
-        print("File 'reports_record.txt' not found. \nAdd the reports from the HRCLM exported to notepad, then run again.")
-        f = open("reports_record.txt","w")
+        print("File 'reports_record.dat' not found. \nAdd the reports from the HRCLM exported to notepad, then run again.")
+        f = open("reports_record.dat","w")
         f.close()
         sys.exit()
     
-    report_list = f.readlines()
+    report_list = f.readlines()[1:]
     
     for i in range(len(report_list)):
         report_list[i] = report_list[i].strip("\n")
@@ -30,6 +30,7 @@ def open_reports():
                 if len(report_list[i][3][d]) == 2:
                     report_list[i][3][d] = "20" + report_list[i][3][d]
             report_list[i][3][d] = int(report_list[i][3][d])
+        report_list[i][3][0],report_list[i][3][2] = report_list[i][3][2],report_list[i][3][0]
     return report_list
         
 def _get_names(HRC_list):
@@ -43,6 +44,9 @@ def _get_dates(HRC_list):
     date_list = ["/".join(date) for date in date_list]
     return date_list
 
+def _clear_records():
+    open("reports_record.dat","w")
+    
 def add_report(rep,IGN,verdict,table):
     conn = sqlite3.connect("HRC_records.db")
     cursor = conn.cursor()
@@ -60,10 +64,14 @@ def indiv_report():
         v = open("reports_to_add.txt",'r')
     except:
         print("'reports_to_add.txt' not found. Please do not remove the file, only edit it! A new one will be automatically created")
+        v = open("reports_to_add.txt",'w') #refresh file, v.truncate() isn't working?
+        v.write("#This file is for manually adding a report to the system.\n#The report structure is as follows:\n#Reporter IGN verdict day-month-year ; mind the spacing and hyphens! Use numbers for date only")
+        return 0
         
     new_indiv = v.readlines()[3:]
     
-    v = open("reports_to_add.txt",'w') #refresh file, v.truncate() isn't working?
-    v.write("#This file is for manually adding a report to the system.\n#The report structure is as follows:\n#Reporter IGN verdict day-month-year ; mind the spacing and hyphens! Use numbers for date only")
+    if v: #v exists, is not emtpy/null from only 4 lines of the file i.e. nothing to add
+        v = open("reports_to_add.txt",'w') #refresh file, v.truncate() isn't working?
+        v.write("#This file is for manually adding a report to the system.\n#The report structure is as follows:\n#Reporter IGN verdict day-month-year ; mind the spacing and hyphens! Use numbers for date only")
 
     return new_indiv
